requirements.txt
# Core dependencies for multilingual-translator
openai-whisper
transformers
sentencepiece
torch
gtts
coqui-tts
soundfile
langdetect
numpy
scipy
pytest
sounddevice   # For recording audio from microphone
pydub
